pyproject.toml
[tool.poetry]
name = "ayureye-gen2"
version = "0.1.0"
description = ""
authors = ["Bishal Neupane <85444310+earthPerson-001@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
djangorestframework = "^3.15.1"
django-debug-toolbar = "^4.3.0"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^7.3.0"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
